added download buttons options for all file formats, still bugs
