+++ packages/editor/src/state/atoms.ts
@@ -10,6 +10,7 @@ import registry from "@penrose/examples/dist/registry.js";
 import { Actions, BorderNode, TabNode } from "flexlayout-react";
 import localforage from "localforage";
 import { debounce, range } from "lodash";
+import { RefObject } from "react";
 import toast from "react-hot-toast";
 import {
   atom,
@@ -292,6 +293,17 @@ export type Diagram = {
   warnings: PenroseWarning[];
   metadata: DiagramMetadata;
 };
+
+export type Canvas = {
+  ref: RefObject<HTMLDivElement> | null;
+};
+
+export const canvasState = atom<Canvas>({
+  key: "canvasState",
+  default: {
+    ref: null,
+  },
+});
 
 export const diagramState = atom<Diagram>({
   key: "diagramState",
